# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tontines/{id}/start": {
      "post": {
        "operationId": "CyclesController_start_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Démarrer maintenant si la date est atteinte et les conditions remplies (admin) — US-4.3",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/start-check": {
      "get": {
        "operationId": "CyclesController_startCheck_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Conditions de démarrage non remplies (admin)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/draw-order": {
      "put": {
        "operationId": "CyclesController_drawOrder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1,
                    "maxItems": 50
                  }
                },
                "required": ["memberIds"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ordre de passage (mode ordre fixe ; après démarrage : cycles futurs uniquement)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/draw-proof": {
      "get": {
        "operationId": "CyclesController_drawProof_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preuve du tirage (graine, hash SHA-256, horodatage, vérification)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/priority-requests": {
      "post": {
        "operationId": "CyclesController_priority_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 1000
                  }
                },
                "required": ["reason"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Demande prioritaire motivée (mode besoin prioritaire, A-15)",
        "tags": ["Tontines — cycles et contributions"]
      },
      "get": {
        "operationId": "CyclesController_priorities_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Demandes prioritaires (admin : toutes ; membre : les siennes)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/cycles/{cycleId}/beneficiary": {
      "post": {
        "operationId": "CyclesController_designate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": ["memberId"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Désigner le bénéficiaire d’un cycle (admin, mode besoin prioritaire)",
        "tags": ["Tontines — cycles et contributions"]
      },
      "get": {
        "operationId": "CyclesController_cycleBeneficiary[0]_v1",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bénéficiaire d’un cycle, preuve de désignation et versement",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/cycles": {
      "get": {
        "operationId": "CyclesController_list_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cycles de la tontine",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/cycles/{cycleId}": {
      "get": {
        "operationId": "CyclesController_one_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d’un cycle (admin : toutes les contributions ; membre : la sienne)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/cycles": {
      "get": {
        "operationId": "CyclesController_listFlat_v1",
        "parameters": [
          {
            "name": "tontineId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cycles d’une tontine (alias de /tontines/{id}/cycles)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/cycles/{cycleId}": {
      "get": {
        "operationId": "CyclesController_oneFlat_v1",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d’un cycle (alias de /tontines/{id}/cycles/{cycleId})",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/cycles/{cycleId}/contributions": {
      "get": {
        "operationId": "CyclesController_cycleContributions[0]_v1",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "not": {}
                },
                {
                  "type": "string",
                  "enum": ["PENDING", "PAID", "LATE", "PAID_LATE", "DEFAULTED"]
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Contributions d’un cycle (admin : toutes ; membre : la sienne), filtre par statut",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/cycles/{cycleId}/contributions": {
      "get": {
        "operationId": "CyclesController_cycleContributions[1]_v1",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "not": {}
                },
                {
                  "type": "string",
                  "enum": ["PENDING", "PAID", "LATE", "PAID_LATE", "DEFAULTED"]
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Contributions d’un cycle (admin : toutes ; membre : la sienne), filtre par statut",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/cycles/{cycleId}/beneficiary": {
      "get": {
        "operationId": "CyclesController_cycleBeneficiary[1]_v1",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bénéficiaire d’un cycle, preuve de désignation et versement",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/contributions/{contributionId}/pay": {
      "post": {
        "operationId": "CyclesController_pay_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contributionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Clé unique (8-128 caractères) par opération",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Payer sa contribution depuis le wallet (hold + capture, pénalité incluse) — US-5.4",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/contributions/{contributionId}/pay": {
      "post": {
        "operationId": "CyclesController_payFlat_v1",
        "parameters": [
          {
            "name": "contributionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Clé unique (8-128 caractères) par opération",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Payer sa contribution (alias de tontines/{id}/contributions/{id}/pay)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/contributions/{contributionId}": {
      "get": {
        "operationId": "CyclesController_contribution_v1",
        "parameters": [
          {
            "name": "contributionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Une contribution (débiteur, admin de la tontine ou super-admin)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/tontines/{id}/entry-fee/pay": {
      "post": {
        "operationId": "CyclesController_entryFee_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Clé unique (8-128 caractères) par opération",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Payer le droit d’entrée (avant démarrage)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
    "/api/v1/me/contributions": {
      "get": {
        "operationId": "CyclesController_mine[0]_v1",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mes échéances (toutes tontines)",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
...
    "/api/v1/tontines/{id}/accounts": {
      "get": {
        "operationId": "CyclesController_accountsList_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comptes de la tontine (US-10.2)",
        "tags": ["Tontines — cycles et contributions"]
      },
      "post": {
        "operationId": "CyclesController_accountsCreate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 100
                  },
                  "type": {
                    "type": "string",
                    "enum": ["MAIN", "SOLIDARITY", "SAVINGS", "LOAN"]
                  },
                  "rules": {
                    "type": "object",
                    "properties": {
                      "interestRatePercent": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 100
                      },
                      "exitConditions": {
                        "type": "string",
                        "maxLength": 1000
                      }
                    },
                    "additionalProperties": false,
                    "default": {}
                  }
                },
                "required": ["name", "type"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un compte (solidarité, épargne, prêt) — configuration uniquement en V1",
        "tags": ["Tontines — cycles et contributions"]
      }
    },
...
  },
  "info": {
    "title": "TontineMoney API",
    "description": "API REST versionnée du MVP TontineMoney (monolithe modulaire). Erreurs au format Problem Details (RFC 9457). Montants en chaînes décimales (unités majeures) + devise ISO 4217. En-tête Idempotency-Key obligatoire sur les POST financiers.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tm_rt"
      }
    },
...
  }
}